refactor orbit-controller operators, webhooks, and controllers
