+++ images/orbit-controller/src/setup.py
@@ -38,6 +38,5 @@ setup(
         "python-jose~=3.2.0",
         "kopf~=1.33.0",
     ],
-    entry_points={"console_scripts": ["orbit-controller = orbit_controller.__main__:main"]},
     include_package_data=True,
 )
